+++ BetterChests/Framework/Services/Features/CategorizeChest.cs
@@ -143,9 +143,9 @@ internal sealed class CategorizeChest : BaseFeature<CategorizeChest>
 
         // Allow transfer if item does matches category
         var itemMatcher = this.GetOrCreateItemMatcher(e.Into);
-        if (!itemMatcher.IsEmpty && itemMatcher.MatchesFilter(e.Item))
+        if (itemMatcher.IsEmpty || !itemMatcher.MatchesFilter(e.Item))
         {
-            e.AllowTransfer();
+            e.PreventTransfer();
         }
     }
 
@@ -156,10 +156,9 @@ internal sealed class CategorizeChest : BaseFeature<CategorizeChest>
             itemMatcher = this.itemMatcherFactory.GetDefault();
         }
 
-        if (itemMatcher.IsEmpty && container.Options.CategorizeChestTags.Any())
-        {
-            itemMatcher.SearchText = string.Join(' ', container.Options.CategorizeChestTags);
-        }
+        itemMatcher.SearchText = container.Options.CategorizeChestTags.Any()
+            ? string.Join(' ', container.Options.CategorizeChestTags)
+            : string.Empty;
 
         this.cachedItemMatchers.AddOrUpdate(container, itemMatcher);
         return itemMatcher;